LayoutTests/inspector/styles-url-linkify.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="../http/tests/inspector/inspector-test2.js"></script>
<script src="elements-tests2.js"></script>
<link rel="stylesheet" href="resources/styles-url-linkify.css">
<script>

function runAfterIframeLoaded()
{
    function step()
    {
        if (!window.iframeLoaded)
            setTimeout(step, 100);
        else
            runTest();
    }
    setTimeout(step, 100);
}

function test()
{
    function dumpHref()
    {
        var href;
        var valueChildNodes = WebInspector.panels.elements.sidebarPanes.styles.sections[0][2].propertiesTreeOutline.children[0].valueElement.childNodes;
        for (var i = 0; i < valueChildNodes.length; ++i) {
            if (valueChildNodes[i].href) {
                href = valueChildNodes[i].href;
                break;
            }
        }
        if (!href) {
            InspectorTest.addResult("href property not found");
            return;
        }
        var segments = href.split("/");
        var output = [];
        for (var i = segments.length - 1, minSegment = i - 3; i >= 0 && i >= minSegment; --i)
            output.unshift(segments[i]);
        InspectorTest.addResult(output.join("/"));
    }

    function selectLocalElement()
    {
        InspectorTest.selectElementAndRun("local", executeLocalTest);
    }

    function completeURL(baseURL, href)
    {
        InspectorTest.addResult(WebInspector.completeURL(baseURL, href));
    }

    function completeURLTest()
    {
        InspectorTest.addResult("Partial URLs completed:");
        completeURL("http://example.com", "/");
        completeURL("http://example.com", "moo");
        completeURL("http://example.com/", "https://secure.com/moo");
        completeURL("https://example.com/foo", "//secure.com/moo");
        completeURL("http://example.com/foo/zoo", "/moo");
        completeURL("http://example.com/foo/zoo/", "moo");
        completeURL("http://example.com/foo/zoo", "boo/moo");
        completeURL("http://example.com/foo", "moo");
        completeURL("http://example.com/foo", "?a=b");
        completeURL("http://example.com/foo?c=d", "?a=b");
    }

    function executeLocalTest()
    {
        completeURLTest();
        InspectorTest.addResult("Link for a URI from CSS document:");
        dumpHref();
        InspectorTest.selectElementAndRun("iframed", executeIframedTest);
    }

    function executeIframedTest()
    {
        InspectorTest.addResult("Link for a URI from iframe inline stylesheet:");
        dumpHref();
        InspectorTest.completeTest();
    }

    InspectorTest.expandDOMSubtreeAndRun(null, selectLocalElement);
}

</script>
</head>
<body onload="runAfterIframeLoaded()">
<p>
Tests that URLs are linked to and completed correctly. Bugs <a href="http://bugs.webkit.org/show_bug.cgi?id=51663">51663</a>, <a href="http://bugs.webkit.org/show_bug.cgi?id=53171">53171</a>
</p>
<div id="local"></div>
<iframe src="resources/styles-url-linkify-iframe.html"></iframe>

</body>
</html>
